backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
langchain==0.2.16
langchain-groq==0.1.9
langgraph==0.1.19
chromadb==0.5.0
pypdf2==3.0.1
python-dotenv==1.0.1
pydantic==2.9.0
aiofiles==24.1.0
langchain-community==0.2.16
langchain-chroma==0.1.4
langchain-huggingface==0.0.2
sentence-transformers==3.3.1
